CI: specify a timeout for the "run app" step
There have been occasions where the run-step got stuck for no apparent
reason, for example:

https://github.com/canonical/flutter-snap/actions/runs/1842641473

According to the output, the app ran fine but for some reason, the step
(Flutter tool?) hung after the app was closed.

Specify a timeout to avoid that the job is stuck for several hours.
15 minutes is a very generous timeout for a step that according to the
logs typically takes less than a minute.