Fix typo causing incorrect `LIBGL_DRIVERS_PATH`
This loop references the nonexistent variable `HOST_DIR_DIRS` when it
should be appending to `HOST_DRIVERS_PATH`. That means it sets
`HOST_DRIVERS_PATH` ends up containing only the last possible directory,
which does not work if that is not the correct one (for example: if it
is the 32-bit directory on a 64-bit multilib system).

I have not tested that the entire snap works correctly with this fix,
but I have manually tested that the loop works, and that the snap works
if I manually set LIBGL_DRIVERS_PATH to include the path this fix adds.

Fixes #93